backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
sqlmodel
pymysql
cryptography
pydantic
python-dotenv
alembic
httpx
python-jose
python-multipart
passlib==1.7.4
bcrypt==4.0.1